// src/new-client-mgmt/enum/bulk-import.enum.ts
export enum BulkImportType {
    MenuItem = 'Menu Item',
    Material = 'Material',
    GLVocuher = 'GL Voucher',
    Expense = 'Expense',
    Purchase = 'Purchase',
    PO = 'Purchase Order',
    GR = 'Goods Receipt',
    IR = 'Invoice Receipt',
    ProductionEvent = 'Production Event',
    InventoryCount = 'Inventory Count',
    WasteEvent = 'Waste Event',
    StockTransfer = 'Stock Transfer',
    Inventory = 'Inventory',
    InventoryHistory = 'Inventory History',
    VendorInvoice = 'Vendor Invoice',
    VendorPayment = 'Vendor Payment',
    CustomerInvoice = 'Customer Invoice',
    CustomerPayment = 'Customer Payment',
    AssetMaster = 'Asset Master',
    AssetAcq = 'Asset Acquisition',
    AssetDep = 'Asset Depreciation',
    AssetRet = 'Asset Retirement'
}